requirements.txt
python-dateutil
pandas
matplotlib
python-dotenv
requests